HF_DEPLOY/requirements.txt
pyTelegramBotAPI==4.20.0
python-dotenv==1.0.1
apscheduler==3.10.4
Flask==3.0.2
gunicorn==21.2.0
requests==2.31.0
